% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilter.r
\name{prefilter}
\alias{prefilter}
\title{Prefilter}
\usage{
prefilter(dat, metadata, min_obs, min_days, vmax, sp = NULL,
  path2repo = NULL)
}
\arguments{
\item{dat}{provide example royal penguin tracking data}

\item{metadata}{provide example royal penguin metadata}

\item{min_obs}{minimum number of observation records a deployment must have}

\item{min_days}{minimum number days a deployment must last}

\item{vmax}{maximum travel rate (m/s) for speed filter}

\item{sp}{for future functionality, not currently used}

\item{path2repo}{for future functionality, not currently used}
}
\value{
writes a tibble to a .csv file with the following variables:
\item{\code{id}}{individual animal id}
\item{\code{date}}{POSIX date-time}
\item{\code{lc}}{location quality class}
\item{\code{lon}}{longitude}
\item{\code{lat}}{latitude}
\item{\code{device_type}}{tag type: GPS, PTT, or GLS}
\item{\code{keep}}{should record be used of ignored by SSM filter (logical)}
}
\description{
Prefilter
}
\details{
Prepare standardised track data for state-space filtering by performing
further standardisation and culling unacceptably short deployments

The track data are fetched automatically from the RAATD repository by specifying
a species' abbreviated name.
}
\examples{
\dontrun{
## prefilter RAATD royal penguin tracks
## load example data
data(rope)

pfd <- prefilter(
  dat = rope,
  metadata = meta,
  min_obs = 30,
  min_days = 5,
  vmax = 10
  )
}

}
