% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_sfilter.r
\name{redo_sfilter}
\alias{redo_sfilter}
\title{redo_sfilter}
\usage{
redo_sfilter(ssm_obj, data, s.inc = 0.1, n.inc = 1, tries = 5,
  common.tau = FALSE)
}
\description{
redo_sfilter
}
\details{
Internal function not normally called by user

redo ssm filter for tracks that failed to converge,
 trying incrementally different \code{span} & \code{nu} values.
 Up to \code{tries} re-filter attemps are made. If \code{common.tau = TRUE},
 function then searches for cases where hat(gamma) > 0.9, deployment < 30 d,
 device_type == PTT & refits sfilter. This may reduce over-smoothing of these
 short & typically sparsely observed tracks
}
