% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfilter.r
\name{sfilter}
\alias{sfilter}
\title{SSM filter}
\usage{
sfilter(d, ts = list(gps = 1, gls = 12, ptt = 2), ...)
}
\arguments{
\item{d}{input data from \code{prefilter} as a tibble of individual tracks grouped by id}

\item{ts}{specify a list of time steps (in h) for gps, gls, & ptt datasets}

\item{...}{additional arguments passed to \code{ssmTMB::fit_ssm}}
}
\value{
a list with 10 elements (see ?ssmTMB::fit_ssm)
}
\description{
State-space filter
}
\details{
Wrapper function that takes data prepared by \code{prefilter}, assigns appropriate
time steps for GPS, GLS and/or PTT devices & calls \code{ssmTMB::fit_ssm} to do the
filtering. \code{ssmTMB} can be installed via \code{devtools::install_github("ianjonsen/ssmTMB")}.
}
\examples{
\dontrun{
## run prefilter with local copies of data and but do not write output to files
pfd <- prefilter(
  sp = "ROPE",
  min_obs = 30,
  min_days = 5,
  vmax = 10,
  fullpath = c(file.path("data","metadata.csv"), file.path("data","rope.csv"), NA, NA)
  )

ssm_by_id <- pfd \%>\%
  do(ssm = sfilter(., span = 0.4, nu = 5))

## try re-filtering tracks that failed to converge
ssm_by_id <- redo_sfilter(ssm_by_id, pfd, tries = 10)

## generate QC plots
ssm_by_id \%>\% qc_plot(sp = "rope", fullpath = file.path(getwd()))
}

}
