% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.r
\name{qc_plot}
\alias{qc_plot}
\title{qc_plot}
\usage{
qc_plot(ssm.tbl, sp, filt_rnd = 1, fullpath = NULL)
}
\arguments{
\item{ssm.tbl}{a compound tibble of SSM filter output grouped by \code{id} or
\code{id} and \code{stage}}

\item{sp}{the species' 4-letter abbreviated name}

\item{filt_rnd}{the filtering round for which plots are generated (one of: 1, 1b, 2, 2b)}

\item{fullpath}{user-provided full path for .pdf file (default = NULL)}
}
\value{
a .pdf file of quality control plots
}
\description{
Generate quality control plots
}
\details{
generate a .pdf of SSM fits to data for qc
}
\examples{
\dontrun{
ssm_by_id \%>\% qc_plot(sp = "soes")
}
}
