% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.r
\name{qc_plot}
\alias{qc_plot}
\title{qc_plot}
\usage{
qc_plot(ssm.tbl, sp, filt_rnd = 1)
}
\arguments{
\item{ssm.tbl}{a compound tibble of SSM filter output grouped by \code{id} or
\code{id} and \code{stage}}

\item{sp}{the species' 4-letter abbreviated name}

\item{filt_rnd}{the filtering round for which plots are generated (one of: 1, 1b, 2, 3)}
}
\value{
a .pdf file of quality control plots
}
\description{
Generate quality control plots
}
\details{
Prepare standardised track data for state-space filtering by performing
further standardisation and culling unacceptably short deployments

The track data are fetched automatically from the RAATD repository by specifying
a species' abbreviated name.
}
\examples{
\dontrun{
ssm_by_id \%>\% qc_plot(sp = "soes")
}
}
