% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_sfilter.r
\name{redo_sfilter}
\alias{redo_sfilter}
\title{redo_sfilter}
\usage{
redo_sfilter(ssm_obj, data, s.inc = 0.1, n.inc = 1, tries = 5,
  common.tau = FALSE)
}
\arguments{
\item{ssm_obj}{output from \code{sfilter()} - a \code{tibble} grouped by id or is and stage}

\item{data}{data object provided by \code{prefilter()}}

\item{s.inc}{increment for \code{span} argument}

\item{n.inc}{increment for \code{nu} argument}

\item{tries}{number of attempts to successfully filter all convergence failures before giving up}

\item{common.tau}{(logical) fit SSM with a single SD for measurement error in both lon & lat}
}
\description{
\code{redo_sfilter()} iteratively re-fits the state-space model to tracks
that have failed to converge. It iteratively adjusts the values for
\code{span} and \code{nu} and makes \code{tries = n} attempts before giving up.
}
\details{
redo ssm filter for tracks that failed to converge,
 trying incrementally different \code{span} & \code{nu} values.
 Up to \code{tries} re-filter attemps are made. If \code{common.tau = TRUE},
 function then searches for cases where hat(gamma) > 0.9, deployment < 30 d,
 device_type == PTT & refits sfilter. This may reduce over-smoothing of these
 short & typically sparsely observed tracks
}
