% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfilter.r
\name{sfilter}
\alias{sfilter}
\title{sfilter}
\usage{
sfilter(d, ts = list(gps = 1, gls = 12, ptt = 2), ...)
}
\arguments{
\item{d}{input data from \code{prefilter} as a tibble of individual tracks grouped by id}

\item{ts}{specify a list of time steps (in h) for gps, gls, & ptt datasets}

\item{...}{additional arguments passed to \code{ssmTMB::fit_ssm}}
}
\value{
a list with 10 elements (see ??fit_ssm)
}
\description{
State-space filter
}
\details{
Wrapper function that takes data prepared by \code{prefilter}, assigns appropriate
time steps for GPS, GLS and/or PTT devices & calls \code{ssmTMB::fit_ssm} to do the
filtering. \code{ssmTMB} can be installed via \code{devtools::install_github("ianjonsen/ssmTMB")}.
}
\examples{
\dontrun{
pfd <- prefilter(
  sp = "SOES",
  min_obs = 30,
  min_days = 5,
  vmax = 10,
  path2data = "~/Dropbox/r"
  )

ssm_by_id <- pfd \%>\%
  dplyr::do(ssm = sfilter(., span = 0.4, nu = 5))

## try re-filtering tracks that failed to converge
ssm_by_id <- redo_sfilter(ssm_by_id, pfd, tries = 10)
}

}
