% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfilter.r
\name{sfilter}
\alias{sfilter}
\title{State-space filter RAATD tracking data}
\usage{
sfilter(d, ts = list(gps = 1, gls = 12, ptt = 2), ...)
}
\arguments{
\item{d}{input data from \code{prefilter()} as a tibble of individual tracks grouped by id}

\item{ts}{specify a list of time steps (in h) for gps, gls, & ptt datasets}

\item{...}{additional arguments passed to \code{ssmTMB::fit_ssm()}. Two common arguments are
\code{span}, the degree of loess smoothing used to estimate initial values for the location states, and
\code{nu}, the degree of freedom to be used in the t-distributed error models for lon & lat}
}
\value{
a list with 10 elements (see \code{?ssmTMB::fit_ssm()})
}
\description{
\code{sfilter()} fits a simple state-space model to fit to pre-filtered RAATD data and returns output
 as a tibble grouped by individual id or individual id and breeding stage.
}
\details{
Wrapper function that takes data prepared by \code{prefilter()}, assigns appropriate
time steps for GPS, GLS and/or PTT devices & calls \code{ssmTMB::fit_ssm()} to do the
filtering. \code{ssmTMB} can be installed via \code{devtools::install_github("ianjonsen/ssmTMB")}.
Note, as \code{sfilter()} imports functions from the \code{ssmTMB} package, you must ensure you
have installed the \code{TMB} package and it's dependencies
}
\examples{
\dontrun{
## run prefilter with example royal penguin data & metadata
data(rope)

pfd <- prefilter(
  dat = rope,
  metadata = meta,
  sp = "ROPE",
  min_obs = 30,
  min_days = 5,
  vmax = 10
  )

ssm_by_id <- pfd \%>\%
  do(ssm = sfilter(., span = 0.4, nu = 5))

## re-filter tracks that failed to converge
ssm_by_id <- redo_sfilter(ssm_by_id, pfd, tries = 10)

## generate QC plots
ssm_by_id \%>\% qc_plot(sp = "rope")
}

}
